Alterando o "int" por "float" :D

+++ D013.py
@@ -1,4 +1,4 @@
 # Faça um algoritmo que leia o salário de um funcionário e mostre seu novo salário, com 15% de aumento.
 print('================== Desafio 13 ======================')
-sal = int(input('Informe o salário: '))
+sal = float(input('Informe o salário: '))
 print(f'O salário com acréscimo de 15% é de {sal+(sal/100*15)}')
